Modify the schedule config format to have a zone group that gets reused, instead of a seperate list in each zone-rule. Fix the not active heartbeat logging to only print on the half-hour stamps. Various compile and runtime cleanups.

+++ src/Switch.cpp
@@ -72,7 +72,7 @@ SwitchLogEntry::getEpochTimeString()
     char tmpStr[64];
     std::string rtnStr;
 
-    sprintf(tmpStr, "%d", epochTime);
+    sprintf(tmpStr, "%d", (int)epochTime);
 
     rtnStr = tmpStr;
